moved Phone, Message and Folder classes to util for clarity, fixed XMLParser not to b greedy, initializes data methods but does not call them upon login
